# Conflicts:
#	pom.xml
#	src/main/java/am/aca/wftartproject/dao/ItemDao.java
#	src/main/java/am/aca/wftartproject/dao/PurchaseHistoryDao.java
#	src/main/java/am/aca/wftartproject/dao/ShoppingCardDao.java
#	src/main/java/am/aca/wftartproject/dao/UserDao.java
#	src/main/java/am/aca/wftartproject/dao/impl/ArtistDaoImpl.java
#	src/main/java/am/aca/wftartproject/dao/impl/ItemDaoImpl.java
#	src/main/java/am/aca/wftartproject/dao/impl/PurchaseHistoryDaoImpl.java
#	src/main/java/am/aca/wftartproject/dao/impl/ShoppingCardDaoImpl.java
#	src/main/java/am/aca/wftartproject/dao/impl/UserDaoImpl.java
#	src/main/java/am/aca/wftartproject/exception/DAOException.java
#	src/main/java/am/aca/wftartproject/exception/ServiceException.java
#	src/main/java/am/aca/wftartproject/model/AbstractUser.java
#	src/main/java/am/aca/wftartproject/model/Artist.java
#	src/main/java/am/aca/wftartproject/model/ArtistSpecialization.java
#	src/main/java/am/aca/wftartproject/model/Item.java
#	src/main/java/am/aca/wftartproject/model/ItemType.java
#	src/main/java/am/aca/wftartproject/model/PurchaseHistory.java
#	src/main/java/am/aca/wftartproject/model/ShoppingCard.java
#	src/main/java/am/aca/wftartproject/model/User.java
#	src/main/java/am/aca/wftartproject/service/ArtistService.java
#	src/main/java/am/aca/wftartproject/service/ItemService.java
#	src/main/java/am/aca/wftartproject/service/PurchaseHistoryService.java
#	src/main/java/am/aca/wftartproject/service/UserService.java
#	src/main/java/am/aca/wftartproject/service/impl/ArtistServiceImpl.java
#	src/main/java/am/aca/wftartproject/service/impl/ItemServiceImpl.java
#	src/main/java/am/aca/wftartproject/service/impl/PurchaseHistoryServiceImpl.java
#	src/main/java/am/aca/wftartproject/service/impl/UserServiceImpl.java
#	src/main/java/am/aca/wftartproject/util/ConnectionPool.java
#	src/main/resources/log4j.properties

+++ src/main/java/am/aca/wftartproject/service/impl/ArtistServiceImpl.java
@@ -2,27 +2,30 @@ package am.aca.wftartproject.service.impl;
 
 import am.aca.wftartproject.dao.ArtistDao;
 import am.aca.wftartproject.dao.impl.ArtistDaoImpl;
+import am.aca.wftartproject.exception.DAOException;
+import am.aca.wftartproject.exception.ServiceException;
 import am.aca.wftartproject.model.Artist;
 import am.aca.wftartproject.service.ArtistService;
+import am.aca.wftartproject.util.DBConnection;
 import org.apache.log4j.Logger;
 
 import java.sql.Connection;
 import java.sql.SQLException;
 
+import static am.aca.wftartproject.service.impl.validator.ValidatorUtil.isEmptyString;
+
 /**
- * Created by ASUS on 30-May-17
+ * Created by surik on 6/3/17
  */
 public class ArtistServiceImpl implements ArtistService {
 
-    private static final Logger LOGGER = Logger.getLogger(ArtistDaoImpl.class);
-    private Connection conn = null;
-    private ArtistDao artistDao = null;
+    private static final Logger LOGGER = Logger.getLogger(ArtistServiceImpl.class);
+
+    private Connection conn = new DBConnection().getDBConnection(DBConnection.DBType.REAL);
+    private ArtistDao artistDao = new ArtistDaoImpl(conn);
 
     public ArtistServiceImpl() throws SQLException, ClassNotFoundException {
-//        Connection conn = new dbconnection().getConnection(ConnectionModel.BASIC).getProductionDBConnection();
-        artistDao = new ArtistDaoImpl(conn);
     }
-
 
 
     /**
@@ -31,9 +34,22 @@ public class ArtistServiceImpl implements ArtistService {
      */
     @Override
     public void addArtist(Artist artist) {
-        artistDao.addArtist(artist);
-    }
 
+        if (artist == null || !artist.isValidArtist()) {
+            LOGGER.error(String.format("Artist is invalid: %s", artist));
+            throw new ServiceException("Invalid artist");
+        }
+
+        try {
+            artistDao.addArtist(artist);
+        } catch (DAOException e) {
+            String error = "Failed to add Artist: %s";
+            LOGGER.error(String.format(error, e.getMessage()));
+            throw new ServiceException(String.format(error, e.getMessage()));
+        }
+
+
+    }
 
     /**
      * @param id
@@ -42,9 +58,19 @@ public class ArtistServiceImpl implements ArtistService {
      */
     @Override
     public Artist findArtist(Long id) {
-        return artistDao.findArtist(id);
-    }
+        if (id == null || id < 0) {
+            LOGGER.error(String.format("Id is invalid: %s", id));
+            throw new ServiceException("Invalid id");
+        }
 
+        try {
+            return artistDao.findArtist(id);
+        } catch (DAOException e) {
+            String error = "Failed to find Artist: %s";
+            LOGGER.error(String.format(error, e.getMessage()));
+            throw new ServiceException(String.format(error, e.getMessage()));
+        }
+    }
 
     /**
      * @param email
@@ -53,10 +79,19 @@ public class ArtistServiceImpl implements ArtistService {
      */
     @Override
     public Artist findArtist(String email) {
-        return artistDao.findArtist(email);
+        if (isEmptyString(email)) {
+            LOGGER.error(String.format("Email is invalid: %s", email));
+            throw new ServiceException("Invalid email");
+        }
+
+        try {
+            return artistDao.findArtist(email);
+        } catch (DAOException e) {
+            String error = "Failed to find Artist: %s";
+            LOGGER.error(String.format(error, e.getMessage()));
+            throw new ServiceException(String.format(error, e.getMessage()));
+        }
     }
-
-
 
     /**
      * @param id
@@ -65,10 +100,23 @@ public class ArtistServiceImpl implements ArtistService {
      */
     @Override
     public void updateArtist(Long id, Artist artist) {
-        artistDao.updateArtist(id, artist);
+        if (id == null || id < 0){
+            LOGGER.error(String.format("Id is invalid: %s", id));
+            throw new ServiceException("Invalid id");
+        }
+        if (artist == null || !artist.isValidArtist()){
+            LOGGER.error(String.format("Artist is invalid: %s", id));
+            throw new ServiceException("Invalid artist");
+        }
+
+        try {
+            artistDao.updateArtist(id, artist);
+        } catch (DAOException e) {
+            String error = "Failed to update Artist: %s";
+            LOGGER.error(String.format(error, e.getMessage()));
+            throw new ServiceException(String.format(error, e.getMessage()));
+        }
     }
-
-
 
     /**
      * @param id
@@ -76,8 +124,18 @@ public class ArtistServiceImpl implements ArtistService {
      */
     @Override
     public void deleteArtist(Long id) {
-        artistDao.deleteArtist(id);
+        if (id == null || id < 0){
+            LOGGER.error(String.format("Id is invalid: %s", id));
+            throw new ServiceException("Invalid id");
+        }
+
+        try {
+            artistDao.deleteArtist(id);
+//        userDao.deleteUser(id);
+        } catch (DAOException e) {
+            String error = "Failed to delete Artist: %s";
+            LOGGER.error(String.format(error, e.getMessage()));
+            throw new ServiceException(String.format(error, e.getMessage()));
+        }
     }
-
-
 }
